refactor: lexical editor value prop and wraper component for variables
