+++ packages/app-page-builder/src/PageBuilder.tsx
@@ -6,7 +6,6 @@ import {
     createProviderPlugin,
     createComponentPlugin
 } from "@webiny/app-admin";
-import { LexicalEditorPlugin } from "@webiny/lexical-editor-pb-element";
 import { PageBuilderProvider as ContextProvider } from "./contexts/PageBuilder";
 import { ReactComponent as PagesIcon } from "./admin/assets/table_chart-24px.svg";
 import { WebsiteSettings } from "./modules/WebsiteSettings/WebsiteSettings";
@@ -104,7 +103,6 @@ export const PageBuilder: React.FC = () => {
         <Fragment>
             <PageBuilderProviderPlugin />
             <EditorRendererPlugin />
-            <LexicalEditorPlugin />
             <Plugins>
                 <PageBuilderMenu />
                 <WebsiteSettings />
