+++ packages/lexical-editor-pb-element/src/components/LexicalText.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+
+//TODO: add code from PbText and PeText
+export const LexicalText: React.FC = () => {
+    return <div>Lexical</div>;
+};
