+++ packages/lexical-editor-pb-element/src/render/plugins/TextElementRendererPlugin.tsx
@@ -3,16 +3,20 @@ import get from "lodash/get";
 import { createComponentPlugin } from "@webiny/react-composition";
 import TextElement from "@webiny/app-page-builder/render/components/Text";
 import { isValidLexicalData, LexicalHtmlRenderer } from "@webiny/lexical-editor";
-import { EditorStateJSONString } from "@webiny/lexical-editor/types";
+import { LexicalValue } from "@webiny/lexical-editor/types";
 
 const DATA_NAMESPACE = "data.text";
 
-// TODO: leave comment here why this plugin even exists
+/**
+ * @description This plugin compose the Legacy TextElement component to support new lexical content
+ * Note: This component is deprecated and will be completely removed in next version when
+ * Lexical editor will be fully implemented in Webiny.
+ */
 export const TextElementRendererPlugin = createComponentPlugin(TextElement, Original => {
     return function TextElementRendererPlugin({ element, rootClassName }): JSX.Element {
-        const textContent = get(element, `${DATA_NAMESPACE}.data.text`);
+        const textContent = get(element, `${DATA_NAMESPACE}.data.text`) as LexicalValue;
         return isValidLexicalData(textContent) ? (
-            <LexicalHtmlRenderer value={textContent as EditorStateJSONString} />
+            <LexicalHtmlRenderer value={textContent} />
         ) : (
             <Original element={element} rootClassName={rootClassName} />
         );
