+++ packages/lexical-editor/src/components/AddRichTextEditorPlugin.tsx
@@ -1,11 +1,12 @@
 import React, { FC } from "react";
 import { createComponentPlugin } from "@webiny/react-composition";
 import { RichTextEditor } from "~/components/Editor/RichTextEditor";
+import { LexicalValue } from "~/types";
 
 interface AddRichTextEditorProps {
     toolbar: React.ReactNode;
     placeholder?: string;
-    initValue: string | null;
+    value: LexicalValue;
     children?: React.ReactNode;
 }
 
@@ -16,13 +17,12 @@ export const AddRichTextEditorPlugin: FC<AddRichTextEditorProps> = ({
 }) => {
     const RichTextEditorPlugin = React.memo(
         createComponentPlugin(RichTextEditor, Original => {
-            return function RichTextEditorElem({ tag, initValue, value, onChange }): JSX.Element {
+            return function RichTextEditorElem({ tag, value, onChange }): JSX.Element {
                 return (
                     <Original
                         toolbar={toolbar}
                         tag={tag}
                         placeholder={placeholder}
-                        initValue={initValue}
                         value={value}
                         onChange={onChange}
                     >
