+++ packages/lexical-editor/src/components/Editor/HeadingEditor.tsx
@@ -2,20 +2,16 @@ import React from "react";
 import { LinkPlugin } from "@lexical/react/LexicalLinkPlugin";
 import { ClickableLinkPlugin } from "~/plugins/ClickableLinkPlugin/ClickableLinkPlugin";
 import { FloatingLinkEditorPlugin } from "~/plugins/FloatingLinkEditorPlugin/FloatingLinkEditorPlugin";
-import { EditorStateJSONString } from "~/types";
+import { LexicalValue } from "~/types";
 import { HeadingToolbar } from "~/components/Toolbar/HeadingToolbar";
 import { RichTextEditor } from "~/components/Editor/RichTextEditor";
 
 interface HeadingEditorProps {
-    initValue: EditorStateJSONString | null;
-    /*
-     * Use "value" prop to update the state of the editor after initialization.
-     */
-    value?: EditorStateJSONString | null;
-    onChange?: (editorState: EditorStateJSONString) => void;
+    value: LexicalValue | null;
+    onChange?: (value: LexicalValue) => void;
     tag?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6";
     placeholder?: string;
-    onBlur?: (editorState: EditorStateJSONString) => void;
+    onBlur?: (editorState: LexicalValue) => void;
 }
 
 export const HeadingEditor: React.FC<HeadingEditorProps> = ({ tag, placeholder, ...rest }) => {
