+++ packages/lexical-editor/src/index.tsx
@@ -37,7 +37,7 @@ export { AddToolbarAction } from "~/components/AddToolbarAction";
 export { AddRichTextEditorPlugin } from "~/components/AddRichTextEditorPlugin";
 export { AddRichTextEditorNodeType } from "~/components/AddRichTextEditorNodeType";
 // utils
-export { getEmptyEditorStateJSONString } from "~/utils/getEmptyEditorStateJSONString";
+export { generateInitialLexicalValue } from "~/utils/generateInitialLexicalValue";
 export { isValidLexicalData } from "~/utils/isValidLexicalData";
 // types
 export * as types from "./types";
