+++ packages/lexical-editor/src/types.ts
@@ -1,2 +1,2 @@
 export type ToolbarType = "heading" | "paragraph" | string;
-export type EditorStateJSONString = string;
+export type LexicalValue = string;
